Lint, split into multiple files
