+++ Sources/TuistGenerator/Linter/TargetLinter.swift
@@ -173,14 +173,16 @@ class TargetLinter: TargetLinting {
     }
 
     private func lintEntitlementsExist(target: Target) -> [LintingIssue] {
-
         var issues: [LintingIssue] = []
         if let entitlements = target.entitlements,
            case let Entitlements.file(path: path) = entitlements,
            !FileHandler.shared.exists(path)
         {
             issues
-                .append(LintingIssue(reason: "Entitlements file not found at path \(entitlements.path!.pathString)", severity: .error))
+                .append(LintingIssue(
+                    reason: "Entitlements file not found at path \(entitlements.path!.pathString)",
+                    severity: .error
+                ))
         }
         return issues
     }
